requirements.txt
streamlit>=1.27.0
numpy>=1.26.0
opencv-python-headless>=4.8.0
imageio>=2.33.0
Pillow>=10.0.0
streamlit-drawable-canvas>=0.9.3